+++ src/main/kotlin/com/github/oxc/project/oxcintellijplugin/oxlint/settings/OxlintSettingsState.kt
@@ -1,5 +1,6 @@
 package com.github.oxc.project.oxcintellijplugin.oxlint.settings
 
+import com.github.oxc.project.oxcintellijplugin.BinarySource
 import com.github.oxc.project.oxcintellijplugin.ConfigurationMode
 import com.github.oxc.project.oxcintellijplugin.oxlint.OxlintUnusedDisableDirectivesSeverity
 import com.intellij.openapi.components.BaseState
@@ -7,6 +8,12 @@ import com.intellij.util.xml.Attribute
 import com.intellij.util.xmlb.annotations.XMap
 
 class OxlintSettingsState : BaseState() {
+
+    @get:Attribute("binarySource")
+    var binarySource by enum(BinarySource.AUTO)
+
+    @get:Attribute("vitePlusPath")
+    var vitePlusPath by string()
 
     @get:Attribute("binaryPath")
     var binaryPath by string()